+++ HISTORY.md
@@ -11,6 +11,7 @@
     * "scn" extension category for cloned scenes and tutorials
     * new "scn_scale([num])" extension, scales and animates a launched tutorial, reports the scale if given no number or zero
     * new "scn_exit" extension, closes the tutorial dialog and redisplays the scenes in the corral
+    * new "scn_position(pane, [x, y])" extension, positions the tutorial dialog into the specified pane
 * **Notable Changes:**
     * always "normalize" SVGs on import, avoids "cut-off" costume parts
 * **Translation Updates:**
@@ -21,6 +22,7 @@
 * extensions: new "scn_scale([num])" extension, scales and animates a launched tutorial, reports the scale if given no number or zero
 * gui, extensions: new "scn_exit" extension, closes the tutorial dialog and redisplays the scenes in the corral
 * extensions: animate tutorial resize actions based on the (hidden) animation preference
+* extensions: new "scn_position(pane, [x, y])" extension, positions the tutorial dialog into the specified pane
 
 ### 2025-12-03
 * gui: always normalize SVGs on import